Moving the CombinedModbusRegister interface to the correct package
