+++ src/main/java/com/github/mdevloo/cube/modbus/communication/conf/SerialModbusCommunication.java
@@ -13,7 +13,7 @@ import com.github.mdevloo.cube.modbus.communication.exception.ModbusReadExceptio
 import com.github.mdevloo.cube.modbus.communication.result.CombinedModbusResult;
 import com.github.mdevloo.cube.modbus.communication.result.ModbusResult;
 import com.github.mdevloo.cube.modbus.register.energy.ModbusRegister;
-import com.github.mdevloo.cube.modbus.register.energy.energy.CombinedModbusRegister;
+import com.github.mdevloo.cube.modbus.register.energy.CombinedModbusRegister;
 import com.github.mdevloo.cube.modbus.service.ModbusService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
